Add files via upload
This my first PowerBi visualization for a data science project, about the what are the stores which you should buy food from if an apocalypse occurs, with Costco being the cheapest one. The visualization also interprets the data that way you will know what foods are right to buy during the apocalypse from the stores as well.